+++ src/main/java/org/core/model/load/electric/LoadValue.java
@@ -4,6 +4,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 
@@ -25,7 +26,7 @@ public class LoadValue {
     private LocalDateTime datetime;
 
     @Column(name = "load_value", nullable = false)
-    private Double loadValue;
+    private BigDecimal loadValue;
 
     @Column(name = "created_at", nullable = false, updatable = false)
     private Timestamp createdAt;
